Move page.WitnessCode{,s} to actor.WitnessCode{,s}
