+++ internal/page/witness_code.go
@@ -5,14 +5,9 @@ import (
 	"errors"
 	"time"
 
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/actor"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/notify"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/random"
-)
-
-const (
-	witnessCodeExpireAfter  = 15 * time.Minute
-	witnessCodeIgnoreAfter  = 2 * time.Hour
-	witnessCodeRequestAfter = time.Minute
 )
 
 var ErrTooManyWitnessCodeRequests = errors.New("too many witness code requests")
@@ -39,7 +34,7 @@ func (s *WitnessCodeSender) SendToCertificateProvider(ctx context.Context, lpa *
 	}
 
 	code := s.randomCode(4)
-	lpa.CertificateProviderCodes = append(lpa.CertificateProviderCodes, WitnessCode{Code: code, Created: s.now()})
+	lpa.CertificateProviderCodes = append(lpa.CertificateProviderCodes, actor.WitnessCode{Code: code, Created: s.now()})
 
 	_, err := s.notifyClient.Sms(ctx, notify.Sms{
 		PhoneNumber: lpa.CertificateProvider.Mobile,
@@ -63,7 +58,7 @@ func (s *WitnessCodeSender) SendToIndependentWitness(ctx context.Context, lpa *L
 	}
 
 	code := s.randomCode(4)
-	lpa.IndependentWitnessCodes = append(lpa.IndependentWitnessCodes, WitnessCode{Code: code, Created: s.now()})
+	lpa.IndependentWitnessCodes = append(lpa.IndependentWitnessCodes, actor.WitnessCode{Code: code, Created: s.now()})
 
 	_, err := s.notifyClient.Sms(ctx, notify.Sms{
 		PhoneNumber: lpa.IndependentWitness.Mobile,
@@ -79,38 +74,4 @@ func (s *WitnessCodeSender) SendToIndependentWitness(ctx context.Context, lpa *L
 	}
 
 	return s.donorStore.Put(ctx, lpa)
-}
-
-type WitnessCode struct {
-	Code    string
-	Created time.Time
-}
-
-func (w WitnessCode) HasExpired() bool {
-	return w.Created.Add(witnessCodeExpireAfter).Before(time.Now())
-}
-
-type WitnessCodes []WitnessCode
-
-func (ws WitnessCodes) Find(code string) (WitnessCode, bool) {
-	for _, w := range ws {
-		if w.Code == code {
-			if w.Created.Add(witnessCodeIgnoreAfter).Before(time.Now()) {
-				break
-			}
-
-			return w, true
-		}
-	}
-
-	return WitnessCode{}, false
-}
-
-func (ws WitnessCodes) CanRequest(now time.Time) bool {
-	if len(ws) == 0 {
-		return true
-	}
-
-	lastCode := ws[len(ws)-1]
-	return lastCode.Created.Add(witnessCodeRequestAfter).Before(now)
 }
